proper typedefs of integer types

+++ arg_parser.h
@@ -14,10 +14,10 @@
 
 typedef int32_t i32;
 typedef uint32_t u32;
-typedef int32_t i16;
-typedef uint32_t u16;
-typedef int32_t i8;
-typedef uint32_t u8;
+typedef int16_t i16;
+typedef uint16_t u16;
+typedef int8_t i8;
+typedef uint8_t u8;
 typedef float f32;
 
 #define ARRAY_LENGTH(ARR) (sizeof(ARR) / sizeof(ARR[0]))
